api/requirements.txt
# Import the internal layer first
-r requirements-internal.txt

# Add Public Libs (External)
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0

# Required for the Hybrid gRPC Server (api_service.py)
grpcio==1.60.0
protobuf==4.25.1